focus.sh now does not change border width
Previously, focus.sh would not work "out of the box" in terms of properly setting the window border to the correct size. This is in relation to fullscreen.sh calling focus.sh, which would give windows an incorrect border size when restored to their normal size. Fullscreen.sh now handles this behavior.